% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacevalue_list.R
\docType{data}
\name{barrier_rename}
\alias{barrier_rename}
\title{Band membership}
\format{
Each is a tibble with two variables and three observations
}
\usage{
barrier_rename
}
\description{
These data sets describe band members of the Beatles and Rolling Stones. They
are toy data sets that can be displayed in their entirety on a slide (e.g. to
demonstrate a join).
}
\examples{
barrier_rename
purpose_rename
satisfaction_rename
}
\keyword{datasets}
