% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmviz_bar.R
\name{kcmviz_bar}
\alias{kcmviz_bar}
\title{Title}
\usage{
kcmviz_bar(
  data,
  outcome_var = data$prop,
  label_var = data$proplabel,
  cat_var = data$element_var,
  ymin = 0,
  ymax = 100,
  lang = "en",
  main_title = "",
  subtitle = "",
  source_info = "",
  order = "ascend",
  color_scheme = "#FDB71A",
  horiz = TRUE
)
}
\description{
Title
}
