% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmviz_clusterbar.R
\name{kcmviz_clusterbar}
\alias{kcmviz_clusterbar}
\title{Title}
\usage{
kcmviz_clusterbar(
  data,
  element_var = data$element_var,
  outcome_var = data$prop,
  lower_bound = data$prop_low,
  upper_bound = data$prop_upp,
  label_var = data$proplabel,
  groupby_var = data$groupby_var,
  ymin = 0,
  ymax = 100,
  main_title = "",
  source_info = "",
  subtitle = "",
  sort = "",
  horiz = TRUE,
  y_label = "",
  x_label = "",
  color_scheme = c("#D67619", "#264d5e", "#006848", "#4B2884", "#FDB71A", "#784885",
    "#a43d6a"),
  label_size = 4.25,
  text_position = 0.75
)
}
\arguments{
\item{text_position}{}
}
\value{
A nice pretty graph
}
\description{
Title
}
