% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmviz_fonts.R
\name{kcmviz_fonts}
\alias{kcmviz_fonts}
\title{Add Google fonts for visualization}
\usage{
kcmviz_fonts()
}
\description{
This function adds Google fonts for visualization purposes using sysfonts and showtext.
}
\keyword{internal}
