% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmviz_stackedbar.R
\name{kcmviz_stackedbar}
\alias{kcmviz_stackedbar}
\title{This function creates a stacked bar plot}
\usage{
kcmviz_stackedbar(
  data,
  outcome_var = data$prop,
  prop_labels = data$proplabel,
  var_labels = data$element_var,
  value_labels = data$response_category,
  main_title = "",
  subtitle = "",
  source_info = "",
  rev_values = FALSE,
  rev_variables = FALSE,
  hide_small_values = TRUE,
  order_bars = FALSE,
  subtitle_h_just = 0,
  fixed_aspect_ratio = TRUE,
  legendnrow = 1,
  color_scheme = NULL,
  colors = NULL,
  textsize_xaxis = 16,
  textsize_yaxis = 16
)
}
\arguments{
\item{data}{A data frame containing the agreement data.}

\item{outcome_var}{The outcome variable representing the proportion of agreement.}

\item{prop_labels}{Labels for the proportions.}

\item{var_labels}{Labels for the variables.}

\item{value_labels}{Labels for the values.}

\item{main_title}{The main title of the plot.}

\item{subtitle}{The subtitle of the plot.}

\item{source_info}{Information about the data source.}

\item{rev_values}{Logical indicating whether to reverse the order of value labels.}

\item{rev_variables}{Logical indicating whether to reverse the order of variable labels.}

\item{hide_small_values}{Logical indicating whether to hide small values in the plot.}

\item{order_bars}{Logical indicating whether to order the bars in the plot.}

\item{subtitle_h_just}{Horizontal justification of the subtitle.}

\item{fixed_aspect_ratio}{Logical indicating whether to maintain a fixed aspect ratio for the plot.}

\item{color_scheme}{A vector of colors for the plot.}

\item{label_size}{The size of the labels.}

\item{text_position}{The position of the text labels.}
}
\value{
A ggplot object representing the agreement visualization.
}
\description{
This function creates a stacked bar plot
}
