% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacevalues.R
\name{replacevalues}
\alias{replacevalues}
\title{Rename Values in a Column}
\usage{
replacevalues(
  data,
  column,
  pattern_replacements,
  na_action = "keep",
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame that is of interest}

\item{column}{The name of the column to transform (the one where the values are to be replaced with new values).}

\item{pattern_replacements}{A named vector with key:value pairs ex. userlanguage = language, sat_bus = Bus, etc.}

\item{na_action}{A character string specifying the action for NA values:
"keep" (default), "remove", or a specific value for replacement.}

\item{verbose}{Boolean indicating if detailed output should be printed.}
}
\description{
Rename Values in a Column
}
