% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycollapse.R
\name{svycollapse}
\alias{svycollapse}
\title{Collapse survey data and compute proportions}
\usage{
svycollapse(data, depvar = NULL, response = NULL, wgt = NULL, group_var = NULL)
}
\arguments{
\item{data}{A data frame containing the survey data.}

\item{depvar}{The column name representing the dependent variable, if not present in data.}

\item{response}{The column name representing the response variable, if not present in data.}

\item{wgt}{The column name representing the weights, if not present in data.}

\item{group_var}{(Optional) The column name representing the grouping variable.}
}
\value{
A summarized data frame with proportions of responses.
}
\description{
This function collapses survey data and computes proportions of responses for a given variable.
}
