% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmviz_bar.R
\name{kcmviz_bar}
\alias{kcmviz_bar}
\title{KCM Survey Data - Bar Graph Visualization}
\usage{
kcmviz_bar(
  data,
  prop = data$prop,
  proplabel = data$proplabel,
  element_var = data$element_var,
  ymin = 0,
  ymax = 100,
  main_title = "",
  subtitle = "",
  source_info = "",
  order = "ascend",
  color_scheme = "#FDB71A",
  horiz = TRUE
)
}
\arguments{
\item{data}{This is the datagram that will be used for the visualization. It should include columns for the category,
as well as the proportion (prop) and the nicely labeled proportion (prop_label). Using the "svycollapse" function, you can
create the data frame necessary to create all the variables.}

\item{prop}{This is the proportion of responses that fall in the category. For categorical variables, it'll be the
proportion of respondents that fit in each category (adding to 100\%). For multiple select, you'll need to make sure to
transpose to long form. Then, it'll be the proportion of respondents who selected yes or no.}

\item{proplabel}{This is the same as the proportion, except with formatting and labeling appropriate for graphs.}

\item{element_var}{This is the variable used for the axis -- the descriptive variable.}

\item{ymin}{This is the minimum of the y-axis for the graph}

\item{ymax}{This is the maximum of the y-axis for the graph}

\item{main_title}{This is the main title of the grpah}

\item{subtitle}{This is the subtitle of the graph (i.e indicating a subgroup or more description)}

\item{source_info}{This is the source of the data for the graph. If it is for Q2 of 2023 Rider/Non-Rider - you can
add source_info="Source: Rider/Non-Rider, Q2 2023"}

\item{order}{This is specifying whether the order of the bars are in ascending or descending order.}

\item{color_scheme}{Default color = #FDB71A}

\item{horiz}{If horiz=TRUE, the graph bars are horizontal and y-axis = \%. If horiz= FALSE, then graph is vertical with
x axis = \%.}
}
\value{
Pretty graph
}
\description{
This function creates bar graphs for showing the distribution of responses -- both across single-response survey
data questions as well as multiple select responses.
}
